requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
black=22.6.0=py39h06a4308_0
blas=1.0=mkl
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.07.19=h06a4308_0
certifi=2022.6.15=py39h06a4308_0
cffi=1.15.0=py39hd667e15_1
charset-normalizer=2.1.1=pypi_0
click=8.1.3=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
conda=4.14.0=py39h06a4308_0
conda-content-trust=0.1.1=pyhd3eb1b0_0
conda-package-handling=1.8.1=py39h7f8727e_0
cryptography=36.0.0=py39h9ce1e76_0
cudatoolkit=11.3.1=h2bc3f7f_2
cytoolz=0.11.0=py39h27cfd23_0
diffusers=0.2.4=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.8.0=pypi_0
flickrapi=2.4.0=pypi_0
freetype=2.11.0=h70c0345_0
ftfy=6.1.1=pypi_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
huggingface-hub=0.9.1=pypi_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.12.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9e=h7f8727e_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h2818925_1
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
lz4-c=1.9.3=h295c915_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
mypy-extensions=0.4.3=pypi_0
mypy_extensions=0.4.3=py39h06a4308_1
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
numpy=1.23.2=pypi_0
oauthlib=3.2.0=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pypi_0
pathspec=0.9.0=py39h06a4308_0
pillow=9.2.0=pypi_0
pip=21.2.4=py39h06a4308_0
platformdirs=2.5.2=pypi_0
pycosat=0.6.3=py39h27cfd23_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.12=h12debd9_0
pytorch=1.12.1=py3.9_cuda11.3_cudnn8.3.2_0
pytorch-mutex=1.0=cuda
pyyaml=6.0=pypi_0
readline=8.1.2=h7f8727e_1
regex=2022.8.17=pypi_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
requests-toolbelt=0.9.1=pypi_0
ruamel_yaml=0.15.100=py39h27cfd23_0
scipy=1.9.0=pypi_0
setuptools=61.2.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.2=hc218d9a_0
tk=8.6.11=h1ccaba5_0
tokenizers=0.12.1=pypi_0
tomli=2.0.1=py39h06a4308_0
toolz=0.11.2=pyhd3eb1b0_0
torchaudio=0.12.1=py39_cu113
torchvision=0.13.1=py39_cu113
tqdm=4.64.0=pypi_0
transformers=4.21.2=pypi_0
typing_extensions=4.3.0=py39h06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.12=pypi_0
wcwidth=0.2.5=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zipp=3.8.1=pypi_0
zlib=1.2.12=h7f8727e_1
zstd=1.5.2=ha4553b6_0
